[Refactor] 서버 핸들러 리팩토링
1. http method 종류 별 핸들러 객체 구현
2. Server에 response 메세지를 전달할 Controller 객체 구현

Related to #14
