+++ Makefile
@@ -11,7 +11,11 @@ SRCS =	webserve/main.cpp \
 		webserve/Configuration/Block/Block.cpp \
 		webserve/Configuration/LocationBlock/LocationBlock.cpp \
 		webserve/Configuration/ServerBlock/ServerBlock.cpp \
-		webserve/ServerHandler/ServerHandler.cpp \
+		webserve/RequestHandler/ServerHandler/ServerHandler.cpp \
+		webserve/RequestHandler/GetHandler/GetHandler.cpp \
+		webserve/RequestHandler/DeleteHandler/DeleteHandler.cpp \
+		webserve/RequestHandler/PostHandler/PostHandler.cpp \
+		webserve/ServerController/ServerController.cpp \
 		webserve/HttpMessage/HttpMessage.cpp \
 		webserve/HttpMessage/HttpRequestMessage.cpp \
 		webserve/HttpMessage/HttpResponseMessage.cpp \
@@ -31,7 +35,11 @@ HEAD =  -I./webserve/Server \
 		-I./webserve/Configuration/Block \
 		-I./webserve/Configuration/ServerBlock \
 		-I./webserve/Configuration/LocationBlock \
-		-I./webserve/ServerHandler \
+		-I./webserve/RequestHandler/ServerHandler \
+		-I./webserve/RequestHandler/GetHandler \
+		-I./webserve/RequestHandler/PostHandler \
+		-I./webserve/RequestHandler/DeleteHandler \
+		-I./webserve/ServerController/ \
 		-I./webserve/HttpMessage \
 		-I./webserve/lib/ft \
 		-I./webserve/CommonLogFormat
