--- webserve/ServerHandler/ServerHandler.cpp
@@ -1,408 +1,0 @@
-#include "ServerHandler.hpp"
-#include <stdio.h>
-#include <stdlib.h>
-
-ServerHandler::ServerHandler() : _request_message()
-{
-    init_status();
-}
-
-ServerHandler::ServerHandler(ServerBlock* server_block, std::string http_message)
-{
-    init_status();
-    _request_message = HttpRequestMessage(http_message);
-    LocationBlock* location_block = findLocationBlock(server_block->getInnerBlock(), 
-                                                      _request_message.getStartLine().getRequestTarget());
-    _config = ConfigDto(*server_block, *location_block);
-}
-
-ServerHandler::~ServerHandler()
-{
-
-}
-
-void ServerHandler::init_status()
-{
-    _status[100] = "Continue";
-    _status[101] = "Switching Protocols";
-    _status[200] = "OK";
-    _status[201] = "Created";
-    _status[202] = "Accepted";
-    _status[203] = "Non-Authoritative Information";
-    _status[204] = "No Content";
-    _status[205] = "Reset Content";
-    _status[206] = "Partial Content";
-    _status[300] = "Multiple Choices";
-    _status[301] = "Moved Permanently";
-    _status[302] = "Found";
-    _status[303] = "See Other";
-    _status[304] = "Not Modified";
-    _status[305] = "Use Proxy";
-    _status[306] = "(Unused)";
-    _status[307] = "Temporary Redirect";
-    _status[308] = "Permanent Redirect";
-    _status[400] = "Bad Request";
-    _status[401] = "Unauthorized";
-    _status[402] = "Payment Required";
-    _status[403] = "Forbidden";
-    _status[404] = "Not Found";
-    _status[405] = "Method Not Allowed";
-    _status[406] = "Not Acceptable";
-    _status[407] = "Proxy Authentication Required";
-    _status[408] = "Request Timeout";
-    _status[409] = "Conflict";
-    _status[410] = "Gone";
-    _status[411] = "Length Required";
-    _status[412] = "Precondition Failed";
-    _status[413] = "Content Too Large";
-    _status[414] = "URI Too Long";
-    _status[415] = "Unsupported Media Type";
-    _status[416] = "Range Not Satisfiable";
-    _status[417] = "Expectation Failed";
-    _status[418] = "(Unused)";
-    _status[421] = "Misdirected Request";
-    _status[422] = "Unprocessable Content";
-    _status[426] = "Upgrade Required";
-    _status[500] = "Internal Server Error";
-    _status[501] = "Not Implemented";
-    _status[502] = "Bad Gateway";
-    _status[503] = "Service Unavailable";
-    _status[504] = "Gateway Timeout";
-    _status[505] = "HTTP Version Not Supported";
-}
-
-HttpRequestMessage& ServerHandler::getRequestMessage(void)
-{
-    return (_request_message);
-}
-
-HttpResponseMessage ServerHandler::getResponseMessage(int status_code, std::string message_body)
-{
-    StatusLine start_line = StatusLine(_request_message.getStartLine().getHttpVersion(), status_code, _status[status_code]);
-    std::map<std::string, std::string> headers;
-    HttpResponseMessage response_message = HttpResponseMessage(start_line, headers, message_body);
-    return (response_message);
-}
-
-bool ServerHandler::checkFile(std::string request_target)
-{
-    int len = request_target.length();
-    if (len < 4)
-        return (true);
-    if (request_target.substr(len - 4, 3).compare("php") == 0)
-        return (false);
-    else if (request_target.substr(len - 4, 3).compare("cgi") == 0)
-        return (false);
-    else if (request_target.substr(len - 3, 2).compare("pl") == 0)
-        return (false);
-    else
-        return (true);
-}
-
-std::vector<std::string> ServerHandler::getIndexPath(std::string root, std::string index)
-{
-    std::vector<std::string> index_path;
-    std::vector<std::string> indexs;
-
-    if (index.empty())
-        return (index_path);
-
-    indexs = ft::splitString(index, " ");
-    std::vector<std::string>::iterator it = indexs.begin();
-    for (; it != indexs.end(); it++)
-    {
-        index_path.push_back(root + "/" + *it);
-    }
-    return (index_path);
-}
-
-LocationBlock* ServerHandler::findLocationBlock(std::vector<Block*> locations, std::string request_target)
-{
-    std::vector<Block*>::iterator it = locations.begin();
-    LocationBlock* location;
-    for (; it != locations.end(); it++)
-    {
-        location = (LocationBlock *)(*it);
-        std::string url = location->getUrl();
-        if (request_target.find(url) == 0)
-            break ;
-    }
-    return (location);
-}
-
-bool ServerHandler::checkAllowMethod(std::string method)
-{
-    std::vector<std::string> config_method = _config.getAllowMethod();
-
-    if (std::find(config_method.begin(), config_method.end(), method) == config_method.end())
-        return (true);
-    return (true);
-}
-
-
-bool ServerHandler::checkDirectory(std::string path)
-{
-    struct stat buf;
-
-    stat(path.c_str(), &buf);
-    if(S_ISDIR(buf.st_mode))
-        return (true);
-    return (false);
-}
-
-std::string ServerHandler::getErrorPage(int status_code)
-{
-    std::string message_body;
-    switch (status_code)
-    {
-    case 404:
-        /* code */
-        message_body = ft::readFileIntoString("var/html/error-page/404.html");
-        break;
-    case 405:
-        /* code */
-        message_body = ft::readFileIntoString("var/html/error-page/405.html");
-        break;
-    case 500:
-        /* code */
-        message_body = ft::readFileIntoString("var/html/error-page/500.html");
-        break;
-    default:
-        message_body = ft::readFileIntoString("var/html/error-page/404.html");
-        break;
-    }
-    return (message_body);
-}
-
-std::string ServerHandler::getAutoIndexPage(std::string path, std::string request_target)
-{
-    std::string message;
-    try
-    {
-        message = "<!DOCTYPE html><html><head><title>";
-        message += request_target + "</title></head><body>";
-        message += "<h1>index of "+ request_target;
-        if (request_target.back() != '/')
-            message += "/";
-        message += "</h1><hr />";
-        message += getDirectoryList(path);
-        message += "<hr /></body></html>";
-    }
-    catch(const Error503Exceptnion& e)
-    {
-        throw ;
-    }
-    return (message);
-}
-
-std::string ServerHandler::getDirectoryList(std::string path)
-{
-    std::stringstream ss;
-    std::string file_name;
-    std::string file_path;
-    DIR *dp = NULL;
-    struct dirent* entry = NULL;
-
-    if ((dp = opendir(path.c_str())) == NULL)
-        throw Error503Exceptnion();
-    ss << "<table>";
-    while ((entry = readdir(dp)) != NULL)
-    {
-        struct stat buf;
-        ss << "<tr>";
-        if (std::strcmp(entry->d_name, ".") == 0)
-            continue ;
-        // file 이름
-        file_name = std::string(entry->d_name);
-        file_path = path;
-        if (file_path.back() != '/')
-            file_path += "/";
-        file_path += file_name;
-        if (checkDirectory(file_path))
-            file_name += "/";
-        ss << "<td width=\"300\"><a href=" << file_name << ">" << file_name << "</a></td>";
-        if (std::strcmp(entry->d_name, "..") == 0)
-        {
-            ss << "<td width=\"200\"></td>";
-            ss << "<td width=\"100\"></td></tr>";
-            continue ;
-        }
-        lstat(file_path.c_str(), &buf);
-        // file 수정 시간
-        ss << "<td width=\"200\">" << std::ctime(&buf.st_ctimespec.tv_sec) << "</td>";
-        // file 크기
-        ss << "<td width=\"100\">" << buf.st_size << "</td>"; 
-        ss << "</tr>";
-    }
-    ss << "</table>";
-    return (ss.str());
-}
-
-std::string ServerHandler::findPath(std::string request_target)
-{
-    std::string root = _config.getRoot();
-    std::string index = _config.getIndex();
-
-    // 요청 url
-    std::string path = root + request_target;
-
-    // 인덱싱 url
-    std::vector<std::string> index_path = getIndexPath(root, index);
-
-    // 디렉토리일 경우
-    if (checkDirectory(path))
-    {
-        // autoindex on 일경우
-        if (_config.getAutoindex())
-            throw AutoIndexExceptnion();
-    }
-    else // 파일일 경우
-    {
-        if (access(path.c_str(), F_OK) == -1)
-        {
-            std::vector<std::string>::iterator it = index_path.begin();
-            for (; it != index_path.end(); it++)
-            {
-                if (access((*it).c_str(), R_OK) == 0)
-                {
-                    return (*it);
-                }
-                else if (access((*it).c_str(), F_OK) == 0)
-                {
-                    // 권한이 없을 경우 [500]
-                    throw Error500Exceptnion();
-                }
-            }
-            // 파일이 없을 경우 [404]
-            throw Error404Exceptnion();
-        }
-    }
-    return(path);
-}
-
-std::string ServerHandler::openFile(std::string request_target)
-{
-    std::string file_path = findPath(request_target);
-    return (ft::readFileIntoString(file_path));
-}
-
-std::string ServerHandler::executeCgi(std::string request_target)
-{
-    std::string file_path = findPath(request_target);
-    // fork 후 cgi 실행
-    return ("");
-}
-
-HttpResponseMessage ServerHandler::getHandler()
-{
-    HttpResponseMessage response_message;
-    int status_code = 200;
-    std::string message_body = "GET";
-    std::string request_target = _request_message.getStartLine().getRequestTarget();
-  
-    // 메소드 권한 확인 [405]
-
-    if (!checkAllowMethod("GET"))
-    {
-        status_code = 405;
-        message_body = getErrorPage(status_code);
-        response_message = getResponseMessage(status_code, message_body);
-        return (response_message);
-    }
-
-    // 파일인지 cgi인지 검사
-    try
-    {
-        if (checkFile(request_target))
-            message_body = openFile(request_target);
-        else
-            message_body = executeCgi(request_target);
-        
-        // 응답 생성
-        response_message = getResponseMessage(status_code, message_body);
-    }
-    catch(const AutoIndexExceptnion& e)
-    {
-        try
-        {
-            std::string file_path = _config.getRoot() + request_target;
-            message_body = getAutoIndexPage(file_path, request_target);
-            response_message = getResponseMessage(status_code, message_body);
-        }
-        catch(const Error503Exceptnion& e)
-        {
-            status_code = 503;
-            message_body = getErrorPage(status_code);
-            response_message = getResponseMessage(status_code, message_body);
-        }
-    }
-    catch(const Error404Exceptnion& e)
-    {
-        status_code = 404;
-        message_body = getErrorPage(status_code);
-        response_message = getResponseMessage(status_code, message_body);
-
-    }
-    catch(const Error500Exceptnion& e)
-    {
-        status_code = 500;
-        message_body = getErrorPage(status_code);
-        response_message = getResponseMessage(status_code, message_body);
-    }
-    catch(const Error503Exceptnion& e)
-    {
-        status_code = 503;
-        message_body = getErrorPage(status_code);
-        response_message = getResponseMessage(status_code, message_body);
-    }
-    return (response_message);
-}
-
-HttpResponseMessage ServerHandler::postHandler()
-{
-    int status_code = 200;
-    std::string message_body = "POST";
-
-    HttpResponseMessage response_message = getResponseMessage(status_code, message_body);
-    return (response_message);
-}
-
-HttpResponseMessage ServerHandler::deleteHandler()
-{
-    int status_code = 200;
-    std::string message_body = "DELETE";
-
-    HttpResponseMessage response_message = getResponseMessage(status_code, message_body);
-    return (response_message);
-}
-
-HttpResponseMessage ServerHandler::requestHandler()
-{
-    std::string http_method;
-    HttpResponseMessage response_message;
-
-    http_method = _request_message.getStartLine().getHttpMethod();
-    if (http_method.compare("GET") == 0)
-        response_message = getHandler();
-    if (http_method.compare("POST") == 0)
-        response_message = postHandler();
-    if (http_method.compare("DELETE") == 0)
-        response_message = deleteHandler();
-    return (response_message);
-}
-
-const char* ServerHandler::AutoIndexExceptnion::what() const throw() {
-  return "Not Found";
-}
-const char* ServerHandler::Error404Exceptnion::what() const throw() {
-  return "Not Found";
-}
-const char* ServerHandler::Error405Exceptnion::what() const throw() {
-  return "Method Not Allowed";
-}
-const char* ServerHandler::Error500Exceptnion::what() const throw() {
-  return "Internal Server Error";
-}
-const char* ServerHandler::Error503Exceptnion::what() const throw() {
-  return "Service Unavailable";
-}
